Speed up duplex: encoder bf16, RTF 0.78 → 0.70
The streaming Voxtral encoder was running fp32 by default — at ~14 ms
every frame, second only to the thinker forward (and we'd been
mis-attributing most of it to "Python overhead"). The fine-grained
profiler `scripts/profile_duplex_fine.py` (added here) materializes
deferred MLX work at every section boundary, which exposed it.

Switching the encoder default to bfloat16 (matching the rest of the
model — thinker, talker, code predictor, mimi all bf16/8-bit already)
halves it to ~7 ms with no audio quality regression.

  RTF 0.78 → 0.70 (10% reduction)
  Frame avg 64 ms → 56 ms
  Headroom under 80 ms: 16 ms → 24 ms (~30%)

Verified deterministically over 8 bench runs (range 0.687-0.706,
tight) and an offline run produces coherent transcripts.

Docs updated: roadmap headline table, optimizations row, root README
and docs/README RTF / headroom numbers.

+++ src/raon_mlx/utils/streaming_encoder.py
@@ -78,7 +78,11 @@ class StreamingAudioEncoder:
         encode_frame(pcm: mx.array) -> mx.array of shape [1, T_adapter, 4096]
     """
 
-    def __init__(self, model_path: str, dtype: mx.Dtype = mx.float32) -> None:
+    def __init__(self, model_path: str, dtype: mx.Dtype = mx.bfloat16) -> None:
+        # bfloat16 default: this encoder runs every frame and was 14 ms in
+        # fp32 — about 22% of the duplex frame budget. The conv stem and
+        # transformer layers are bfloat16-safe (matches the rest of the
+        # model). Mel filters stay fp32 internally for numeric stability.
         self._model_path = model_path
         self._dtype = dtype
